refactor: extract the peer-subscription store to a leaf so peerSync and peerSyncReceive stop importing each other (#5919)
peerSync.js and peerSyncReceive.js formed a two-module ESM cycle on the
federation path: peerSync statically imported applyIncomingPush while the
receiver imported findPeerSubscription/subscribePeer back out of peerSync.
In a static cycle whichever module evaluates first sees undefined bindings,
so a module-ordering change could turn peer sync into a boot-time crash.

The subscription CRUD now lives in sharing/peerSubscriptions.js (a leaf over
peerSyncShared/peerTombstoneCursors/peerSyncPush), along with the
background-operation registry and the debounce timer map unsubscribePeer has
to cancel against. Both peerSync and peerSyncReceive import the declaring
module; peerSync keeps re-exporting the surface so outside callers are
unchanged. Drops the now-dead peerSyncShared imports peerSync no longer uses.
